update layout, remove duplicated code
