+++ src/components/Project/index.js
@@ -1,3 +1,3 @@
-import { MyProject as Project } from './project';
+import { MyProject as Project } from './styles';
 
 export { Project };
