+++ src/components/index.js
@@ -1,9 +1,9 @@
-import { Nav } from './nav';
 import { Foot } from './foot';
 import { Logo } from './logo';
 import { Lock } from './lock';
 import { Switch } from './switch';
+import { Layout } from './layout';
 import { Project } from './project';
 import { ScrollTop } from './scroll-top';
 
-export { Nav, Foot, Logo, Lock, Switch, Project, ScrollTop };
+export { Foot, Logo, Lock, Switch, Layout, Project, ScrollTop };
