+++ src/pages/EA1.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { Nav, Foot } from '../components';
 import { Spread } from '../components/switch/styles';
 import QUERIES, {
   Main,
@@ -53,7 +52,6 @@ const EA1 = ({ toggleMode, mode, spread }) => {
   return (
     <Main style={{ position: 'relative' }}>
       <Spread className={spreadClass} />
-      <Nav mode={mode} toggleMode={toggleMode} />
       <Midi style={{ marginBottom: 100, mixBlendMode: 'difference' }}>
         <Heading2
           style={{
@@ -595,8 +593,6 @@ const EA1 = ({ toggleMode, mode, spread }) => {
           </div>
         </FlexGroup>
       </Midi>
-
-      <Foot />
     </Main>
   );
 };
