+++ src/pages/Vuln.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import React, { useState } from 'react';
 
-import { Nav, Foot, Lock } from '../components';
+import { Lock } from '../components';
 import { Spread } from '../components/switch/styles';
 import QUERIES, { Main, Midi, Full, Heading2, Paragraph, Paragraph2 } from '../theming/styles';
 import {
@@ -48,7 +48,6 @@ const VulnPage = ({ toggleMode, mode, spread }) => {
     >
       <Lock isLocked={isLocked} setIsLocked={setIsLocked} />
       <Spread className={spreadClass} />
-      <Nav mode={mode} toggleMode={toggleMode} />
       <Midi style={{ marginBottom: 100, mixBlendMode: 'difference' }}>
         <Heading2
           style={{
@@ -361,7 +360,6 @@ const VulnPage = ({ toggleMode, mode, spread }) => {
           alt="Low fidelity wireframes that display movie theater seat selection flow"
         />
       </Midi>
-      <Foot />
     </Main>
   );
 };
