+++ src/pages/index.js
@@ -3,6 +3,5 @@ import { Movie } from './Movie';
 import { VulnPage } from './Vuln';
 import { Landing } from './Landing';
 import { DevopsPage } from './Devops';
-import { AppWrapper } from './AppWrapper';
 
-export { EA1, Movie, Landing, VulnPage, AppWrapper, DevopsPage };
+export { EA1, Movie, Landing, VulnPage, DevopsPage };
