+++ src/theming/styles.js
@@ -151,10 +151,10 @@ export const Midi = styled.div`
   height: auto;
   margin: 20px auto 50px auto;
   padding: 30px;
-  max-width: 1110px;
+  max-width: 1200px;
 
   @media (${QUERIES.large}) {
-    margin: 100px auto 170px auto;
+    margin: 60px auto 100px auto;
     padding: 0px;
   }
 `;
